Update target build settings to pin Swift Version 2.3
This was (apparently) fixed for earlier Xcode 8 betas by commit
ec2c138, but it doesn't work on Xcode 8 GM (as well as beta8).

So I re-ran the Swift 2.3 conversion on all the framework
and app/test targets that contained Swift code and allowed
Xcode to update the build settings. The difference between GM
and earlier builds seems to be that it is not sufficient to just
set the SWIFT_VERSION on the 'CocoaLumberjackSwift' framework,
but must be done for each Swift target (e.g. the -iOS, -watchOS,
and -tvOS variants).